requirements.txt
flask
pandas
matplotlib
fpdf2
gunicorn
